Fix ts

+++ backend/src/lib.rs
@@ -30,7 +30,7 @@ pub struct Spotify {
 
 #[derive(Clone, Deserialize, Debug, Serialize)]
 pub struct Stats {
-    pub timestamp_ms: u128,
+    pub timestamp_ms: i64,
     pub disk: Disk,
     pub uptime: Uptime,
     pub ram: Ram,
